2007-01-09  Mark Rowe  <mrowe@apple.com>
        Reviewed by Tim H.

        Build fix for 64-bit platforms.

        * platform/graphics/svg/SVGPaintServerGradient.h:
        * platform/graphics/svg/cg/SVGPaintServerGradientCg.cpp:
        (WebCore::SVGPaintServerGradient::updateQuartzGradientStopsCache):


Canonical link: https://commits.webkit.org/15712@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@18695 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/graphics/svg/cg/SVGPaintServerGradientCg.cpp
@@ -129,7 +129,7 @@ void SVGPaintServerGradient::updateQuartzGradientStopsCache(const Vector<SVGGrad
 
     CGFloat previousOffset = 0.0;
     for (unsigned i = 0; i < stops.size(); ++i) {
-        CGFloat currOffset = min(max(stops[i].first, previousOffset), 1.0f);
+        CGFloat currOffset = min(max(stops[i].first, previousOffset), static_cast<CGFloat>(1.0));
         m_stopsCache[i].offset = currOffset;
         m_stopsCache[i].previousDeltaInverse = 1.0 / (currOffset - previousOffset);
         previousOffset = currOffset;
